first-RAG-langchain-project/requirements.txt
langchain
langchain-community
langchain-openai
unstructured
"unstructured[md]"
libmagic
//chromadb
transformers
torch
sentence-transformers
pinecone-client
langchain-pinecone
langchain-huggingface
langchain-ollama
